rightshero/requirements.txt
Django==4.2.16
psycopg2-binary==2.9.6  
mysqlclient
djangorestframework
django-cors-headers

